requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
gunicorn==21.2.0